requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=14.0
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
typer>=0.9
torch>=2.1
faiss-cpu>=1.7.4
lightgbm>=4.0
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
joblib>=1.3
